+++ app/org/sagebionetworks/bridge/play/controllers/BaseController.java
@@ -11,7 +11,6 @@ import java.util.Collection;
 import javax.annotation.Nonnull;
 
 import models.Metrics;
-import models.RequestUtils;
 import models.StatusMessage;
 
 import org.sagebionetworks.bridge.cache.CacheProvider;
@@ -24,6 +23,7 @@ import org.sagebionetworks.bridge.json.BridgeObjectMapper;
 import org.sagebionetworks.bridge.models.accounts.User;
 import org.sagebionetworks.bridge.models.accounts.UserSession;
 import org.sagebionetworks.bridge.models.studies.StudyIdentifier;
+import org.sagebionetworks.bridge.play.interceptors.RequestUtils;
 import org.sagebionetworks.bridge.services.AuthenticationService;
 import org.sagebionetworks.bridge.services.StudyService;
 import org.slf4j.Logger;
