+++ test/org/sagebionetworks/bridge/play/interceptors/RequestUtilsTest.java
@@ -1,4 +1,4 @@
-package models;
+package org.sagebionetworks.bridge.play.interceptors;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -11,6 +11,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 import org.junit.Test;
+import org.sagebionetworks.bridge.play.interceptors.RequestUtils;
 
 import play.mvc.Http.Request;
